auth crud

+++ controllers/books-controller.js
@@ -3,7 +3,7 @@ const Book = require("../model/Book");
 const getAllBooks = async (req, res, next) => {
   let books;
   try {
-    books = await Book.find();
+    books = await Book.find().where("user").equals(req.user.id);
   } catch (err) {
     console.log(err);
   }
@@ -19,6 +19,9 @@ const getById = async (req, res, next) => {
   let book;
   try {
     book = await Book.findById(id);
+    if (book.user.valueOf() !== req.user.id) {
+      return res.status(403).json({ message: "No book" });
+    }
   } catch (err) {
     console.log(err);
   }
@@ -31,14 +34,15 @@ const getById = async (req, res, next) => {
 const addBook = async (req, res, next) => {
   let book;
   try {
-    book = new Book({
+    book = await Book.create({
       ...req.body,
+      user: req.user.id,
     });
+
     await book.save();
   } catch (err) {
     console.log(err);
   }
-
   if (!book) {
     return res.status(500).json({ message: "Unable To Add" });
   }
@@ -49,10 +53,16 @@ const updateBook = async (req, res, next) => {
   const id = req.params.id;
   let book;
   try {
+    book = await Book.findById(id);
+    if (book.user.valueOf() !== req.user.id) {
+      return res.status(403).json({ message: "Unable To Update By this ID" });
+    }
     book = await Book.findByIdAndUpdate(id, {
       ...req.body,
     });
     book = await book.save();
+
+    book = await Book.findById(id);
   } catch (err) {
     console.log(err);
   }
@@ -66,7 +76,11 @@ const deleteBook = async (req, res, next) => {
   const id = req.params.id;
   let book;
   try {
-    book = await Book.findByIdAndRemove(id);
+    book = await Book.findById(id);
+    if (book.user.valueOf() !== req.user.id) {
+      return res.status(403).json({ message: "Unable To Delete By this ID" });
+    }
+    await Book.findByIdAndRemove(id);
   } catch (err) {
     console.log(err);
   }
